[bugfix] out of bounds exception on removeObjectAtIndex, if it is compiled as 64bit binary

+++ pdf2png.m
@@ -28,7 +28,7 @@ int main( int argc, char* argv[] )
 	}
 	
 	// If we have a "--dpi" along with a corresponding argument ...
-	unsigned int index = NSNotFound;
+	NSUInteger index = NSNotFound;
 	if ( (index = [args indexOfObject: @"--dpi"]) != NSNotFound && index + 1 < [args count] )
 	{
 		// Parse it as an integer
